requirements.txt
websockets==14.2
google-genai==0.5.0
markitdown
